Rename program view components
